Refactor : Put our modified methods into a sub-class of Robot
