+++ simulator/src/robot.py
@@ -237,13 +237,7 @@ class Robot:
 
         DO NOT modify robot._linear_velocity or robot._angular_velocity directly. DO NOT modify move()
         """
-        avoid_wall = self.avoid_wall()
-        if not avoid_wall:
-            if dispersion:
-                avoid_robots = self.disperse()
-
-            else:
-                self.flocking()
+        pass
 
     def draw(self, screen):
         # --- IR proximity sensors ---
@@ -288,119 +282,6 @@ class Robot:
         heading_vec = rotate_vector(np.array([self._radius + 2, 0]), self._heading)
         pygame.draw.line(screen, ROBOT_COLOR, self._pos, self._pos + heading_vec, 3)
 
-    def flocking(self):
-        heading_readings = [r.message.heading for r in self.rab_signals]
-        if len(heading_readings) == 0:
-            self.set_rotation_and_speed(0, MAX_SPEED)
-            return False
-
-        # align
-        average_heading = sum(heading_readings) / len(
-            heading_readings
-        )  # Average heading of other robots
-        align_coefficient = self.compute_angle_diff(average_heading)
-
-        # Define some method/metric for too close and should disperse a bit
-        # make delta bearing a sum of separation, cohesion and alignment
-        boids: list[dict] = self._get_boids()
-        # https://math.stackexchange.com/questions/2390443/extracting-x-and-y-values-from-radians
-        centroid = np.mean([boid["angle"] for boid in boids])
-        cohesion_coefficient = centroid
-
-        close_boids = [boid for boid in boids if boid["distance"] < RAB_RANGE / 5]
-
-        separation_coefficient = (
-            0
-            if len(close_boids) == 0
-            else np.deg2rad(
-                180 + np.rad2deg(np.mean([boid["angle"] for boid in close_boids]))
-            )
-        )
-        if separation_coefficient != 0 and self.verbose:
-            print(separation_coefficient)
-
-        delta_bearing = (
-            align_coefficient * 1
-            + separation_coefficient * 1.2
-            + cohesion_coefficient * 0.5
-        )
-        if self.verbose:
-            print(f"Aligning with other robots! Turning {delta_bearing}")
-        self.set_rotation_and_speed(delta_bearing, MAX_SPEED)
-
-        return True
-
-    def avoid(self, angle_readings):
-        average_angle = sum(angle_readings) / len(angle_readings)
-        if average_angle >= 0:  # If wall is to the left or infront
-            opposite_angle = average_angle - math.pi
-        else:  # If wall is to the right
-            opposite_angle = average_angle + math.pi
-
-        # Add randomness
-        random_angle_diff = uniform(math.pi * 0.1, math.pi * -0.1)
-        target_angle = opposite_angle + random_angle_diff
-
-        if self.verbose:
-            print(f"Avoiding! Turning {target_angle}")
-        self.set_rotation_and_speed(target_angle, MAX_SPEED)
-
-    def disperse(self):
-        robot_angles = [
-            r.bearing
-            for r in self.rab_signals
-            if r.distance > self.light_intensity * 100
-        ]  # Distances seem to be around 70 - 150 and light_intensity goes from 0 to 1 it seems.
-        should_disperse = len(robot_angles) > 0
-
-        if should_disperse:
-            self.avoid(robot_angles)
-
-        else:
-            if self.verbose:
-                print("Nothing is wrong:) Stopping and chilling.")
-            self.set_rotation_and_speed(0, 0)
-
-        return should_disperse
-
-    def avoid_wall(self):
-        wall_reading_angles = []
-
-        for i, angle in enumerate(self.prox_angles):
-            if self.prox_readings[i].reading_type == Objects.Wall:
-                if angle > math.pi:
-                    wall_reading_angles.append(angle - 2 * math.pi)
-                else:
-                    wall_reading_angles.append(angle)
-
-        # See if we register any wall not behind us
-        filered_list = [
-            r
-            for r in wall_reading_angles
-            if r != self.prox_angles[2] and r != (self.prox_angles[3] - 2 * math.pi)
-        ]
-        should_activate = len(filered_list) > 0
-
-        if should_activate:
-            self.avoid(wall_reading_angles)
-
-        return should_activate
-
-    def _get_boids(self) -> list[dict[str, float]]:
-        boids = []
-        for sig in self.rab_signals:
-            sig_angle = self._heading + self.rab_angles[sig.sensor_idx]
-            relative_dir = np.array([np.cos(sig_angle), np.sin(sig_angle)])  # radians
-
-            boids.append(
-                {
-                    "distance": sig.distance,
-                    "angle": relative_dir,
-                    "heading": sig.message.heading,
-                }
-            )
-
-        return boids
 
 
 def rotate_vector(vec, angle):
